Fixed header bit field issue in MSVC

+++ include/glaze/binary/header.hpp
@@ -14,19 +14,19 @@ namespace glaze::detail
    static_assert(sizeof(header8) == 1);
    
    struct header16 {
-      uint8_t config : 2;
+      uint16_t config : 2;
       uint16_t size : 14;
    };
    static_assert(sizeof(header16) == 2);
    
    struct header32 {
-      uint8_t config : 2;
+      uint32_t config : 2;
       uint32_t size : 30;
    };
    static_assert(sizeof(header32) == 4);
    
    struct header64 {
-      uint8_t config : 2;
+      uint64_t config : 2;
       uint64_t size : 62;
    };
    static_assert(sizeof(header64) == 8);
